Development (#7)
* init

* - added env file

* - removed loans menu and transfer it under Transactions
- added AddUpdateUserDrawer

* - bug fixes
- added role CRUD
- added permissions CRUD

* added roles and bug fixes

* - added user management
- bug fixes

* - fixed display of branches when editting a user

* - added groups management

* - fixed delete in all modules
- added client management
- bug fixes

* - added loans

* - bug fixes
- added roles for nav and data related
- added details based on the last meeting

* - bug fixes and enhancements

* - added loan applications
- bug fixes
- permissions
and many more...

* - added client detail page
- added modal
- bug fixes

* bug fixes

* updates

* updates

* bug fixes

* - updated the login screen

* - bug fixes

* Stopped tracking .env File

* Stopped tracking next.config.js File

* - added profile picture for clients

* - bug fixes

* - added daily cash collections
- bug fixes

+++ src/pages/api/users/delete.js
@@ -26,7 +26,7 @@ async function deleteUser(req, res) {
             .collection('users')
             .deleteOne({ _id: ObjectId(_id) });
 
-        removeFile.removeFile(user[0].profile);
+        user[0].profile && removeFile.removeFile(user[0].profile);
 
         response = {
             success: true
